requirements.txt
pytest
selenium
webdriver-manager
allure-python-commons